Create Using SSM documents Create Scheduled ECS task
Here's a basic outline of how you could set this up:

1. Create an Automation Document: Create an AWS Systems Manager Automation document that defines the task you want to execute on your ECS cluster. This document can include commands to start tasks, stop tasks, update services, etc.

2.Define a Schedule: Use AWS Scheduler to define a schedule for when you want the Automation document to be executed. You can specify the frequency (e.g., daily, weekly, etc.) and the time of day.

3.Associate Schedule with Automation Document: Configure AWS Scheduler to associate the schedule with the Automation document you created in step 1.


In this example:

We define an AWS Systems Manager Automation document (aws_ssm_document) that will run ECS tasks.
We define an AWS Scheduler schedule (aws_scheduler_schedule) that specifies when the Automation document should be executed. We associate this schedule with the Automation document by specifying its name in the targets block.